All calls of RtAudio methods were wrapped by try/catch, minimalNoteDuration parameter went to double to fix win ussue, MainWindow creates widgets with this parent to avoid flickering under win

+++ src/libs/core/taudioparams.h
@@ -1,5 +1,5 @@
 /***************************************************************************
- *   Copyright (C) 2011-2014 by Tomasz Bojczuk                             *
+ *   Copyright (C) 2011-2015 by Tomasz Bojczuk                             *
  *   tomaszbojczuk@gmail.com                                               *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
@@ -38,7 +38,7 @@ public:
   QString INdevName; // input device name
   float a440diff; // difference between standard a1 440Hz and user preferred base pitch
   float minimalVol; // only above this value detected note is sending to Nootka
-  float minDuration; // minimal duration of a sound above which it is detected
+  qreal minDuration; // minimal duration of a sound above which it is detected
   int detectMethod; // pitch detection method (0 - MPM, 1 - auto-correlation, 2 - MPM modified cepstrum)
 	quint8 intonation; // accuracy of intonation in detected note - corresponds with Eaccuracy
 // duplex mode
